+++ src/main/java/com/limechain/trie/BlockTrieAccessor.java
@@ -1,12 +1,7 @@
 package com.limechain.trie;
 
 import com.limechain.storage.block.BlockState;
-import com.limechain.trie.structure.NodeHandle;
-import com.limechain.trie.structure.database.NodeData;
-import com.limechain.trie.structure.nibble.Nibbles;
 import io.emeraldpay.polkaj.types.Hash256;
-
-import java.util.Optional;
 
 public class BlockTrieAccessor extends TrieAccessor {
 
@@ -16,13 +11,6 @@ public class BlockTrieAccessor extends TrieAccessor {
         if (BlockState.getInstance().isInitialized()) {
             super.lastRoot = BlockState.getInstance().getHeader(blockHash).getStateRoot().getBytes();
         }
-    }
-
-    public Optional<byte[]> findMerkleValue(Nibbles key) {
-        loadPathToKey(key);
-        return partialTrie.existingNode(key)
-                .map(NodeHandle::getUserData)
-                .map(NodeData::getMerkleValue);
     }
 
 }
